utils module for functions and classes
